Update mpi_quad.c
The MPI_Bcast function call specified type MPI_INT to describe the variable "n" which had been initialized as "long int". This causes incorrect results for the integer comparison which controls the for loop when the code is compiled with recent (at least 2020 - 2022) Intel MPI compilers. Correct typing is preferred.

+++ mpi/mpi_quad.c
@@ -126,7 +126,7 @@ int main ( int argc, char *argv[] )
   The master process broadcasts, and the other processes receive,
   the number of intervals N.
 */
-  ierr = MPI_Bcast ( &n, 1, MPI_INT, master, MPI_COMM_WORLD );
+  ierr = MPI_Bcast ( &n, 1, MPI_LONG, master, MPI_COMM_WORLD );
 /*
   Every process integrates F(X) over a subinterval determined by its process ID.
 */
